backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.12
supabase==2.5.1
psycopg[binary]==3.2.3
SQLAlchemy==2.0.35
gitpython==3.1.43
numpy==1.26.4
scikit-learn==1.5.2
orjson==3.10.7
tqdm==4.66.5
# tree-sitter (optional)
tree_sitter==0.21.3
openai==1.3.3
python-dotenv==0.21.1